Add unfinished code for model training
