+++ libs/model/attention_model.py
